- Restructuring package. - Adding story for color picker. - Updated color swatches. - Added a prefix for '#' in the color input.
